requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-filter==23.5
psycopg2-binary==2.9.9
djoser==2.2.0
django-cors-headers==4.3.0
python-dotenv==1.0.0
djangorestframework-simplejwt==5.3.0
setuptools==68.2.2
drf-spectacular==0.24.1
celery==5.3.6
django-celery-beat==2.5.0
redis==5.0.1
pytest==8.2.2
gunicorn==20.1.0
ge==0.0.4